bot/requirements.txt
aiogram==3.1.1
aiohttp==3.9.1
python-decouple==3.8
requests==2.31.0
